refactor: import index  для payments user
