+++ src/modules/users-modules/view-user-info/components/user-payments/components/UserPayments.tsx
@@ -2,36 +2,39 @@ import React, { useMemo, useState } from 'react'
 
 import { useQuery } from '@apollo/client'
 import {
-  flexRender,
   getCoreRowModel,
   useReactTable,
   PaginationState,
   getSortedRowModel,
   getPaginationRowModel,
-  ColumnDef, Table,
+  ColumnDef,
+  Table,
 } from '@tanstack/react-table'
 
 import { capitalizeFirstLetter } from '@/common'
 import {
   dateChangesFormat,
-  setUserPaymentsDataEffect,
-  SkeletonUserPayments,
-} from '@/modules/users-modules/view-user-info'
-import { GET_USER_PAYMENTS } from '@/modules/users-modules/view-user-info/components/user-payments/queries/viewUserPaymentsQueries'
-import {
+  GET_USER_PAYMENTS,
   ItemsPaymentsType,
   PaymentsUser,
+  setUserPaymentsDataEffect,
+  SkeletonUserPayments,
+  UserPaymentsTable,
   UserPaymentsType,
-} from '@/modules/users-modules/view-user-info/components/user-payments/types/UserPaymentsType'
-import {
-  UserPaymentsTable
-} from "@/modules/users-modules/view-user-info/components/user-payments/components/UserPaymentsTable";
+} from '@/modules/users-modules/view-user-info'
 
 export const UserPayments = () => {
   const [{ pageIndex, pageSize }, setPagination] = useState<PaginationState>({
     pageIndex: 0,
     pageSize: 10,
   })
+  const pagination: { pageIndex: number; pageSize: number } = useMemo(
+    () => ({
+      pageIndex,
+      pageSize,
+    }),
+    [pageIndex, pageSize]
+  )
   const { loading, data } = useQuery<UserPaymentsType>(GET_USER_PAYMENTS, {
     variables: {
       userId: Number(10),
@@ -44,50 +47,40 @@ export const UserPayments = () => {
   const pageCount: number | undefined = paymentsUser?.pagesCount
   const [myPaymentsData, setMyPaymentsData] = useState<ItemsPaymentsType[]>([])
 
-  const pagination: { pageIndex: number; pageSize: number } = useMemo(
-    () => ({
-      pageIndex,
-      pageSize,
-    }),
-    [pageIndex, pageSize]
-  )
-
   setUserPaymentsDataEffect(paymentsUser, loading, setMyPaymentsData)
 
-  const columns: ColumnDef<ItemsPaymentsType>[] = useMemo(
-    () => [
-      {
-        header: 'Date of Payment',
-        cell: (params: any) =>
-          loading ? <SkeletonUserPayments /> : dateChangesFormat(params.getValue()),
-        accessorKey: 'dataOfPayment',
-      },
-      {
-        header: 'End date of subscription',
-        cell: (params: any) =>
-          loading ? <SkeletonUserPayments /> : dateChangesFormat(params.getValue()),
-        accessorKey: 'endDateOfSubscription',
-      },
-      {
-        header: 'Price',
-        cell: (params: any) => (loading ? <SkeletonUserPayments /> : '$' + params.getValue()),
-        accessorKey: 'price',
-      },
-      {
-        header: 'Subscription Type',
-        cell: (params: any) =>
-          loading ? <SkeletonUserPayments /> : capitalizeFirstLetter(params.getValue()),
-        accessorKey: 'type',
-      },
-      {
-        header: 'Payment Type',
-        cell: (params: any) =>
-          loading ? <SkeletonUserPayments /> : capitalizeFirstLetter(params.getValue()),
-        accessorKey: 'paymentType',
-      },
-    ],
-    []
-  )
+  const columns: ColumnDef<ItemsPaymentsType>[] = [
+    {
+      header: 'Date of Payment',
+      cell: (params: any) =>
+        loading ? <SkeletonUserPayments /> : dateChangesFormat(params.getValue()),
+      accessorKey: 'dataOfPayment',
+    },
+    {
+      header: 'End date of subscription',
+      cell: (params: any) =>
+        loading ? <SkeletonUserPayments /> : dateChangesFormat(params.getValue()),
+      accessorKey: 'endDateOfSubscription',
+    },
+    {
+      header: 'Price',
+      cell: (params: any) => (loading ? <SkeletonUserPayments /> : '$' + params.getValue()),
+      accessorKey: 'price',
+    },
+    {
+      header: 'Subscription Type',
+      cell: (params: any) =>
+        loading ? <SkeletonUserPayments /> : capitalizeFirstLetter(params.getValue()),
+      accessorKey: 'type',
+    },
+    {
+      header: 'Payment Type',
+      cell: (params: any) =>
+        loading ? <SkeletonUserPayments /> : capitalizeFirstLetter(params.getValue()),
+      accessorKey: 'paymentType',
+    },
+  ]
+
   const tableProps: Table<ItemsPaymentsType> = useReactTable({
     data: myPaymentsData,
     columns: columns,
@@ -102,7 +95,6 @@ export const UserPayments = () => {
     getPaginationRowModel: getPaginationRowModel(),
     getSortedRowModel: getSortedRowModel(),
   })
-console.log('1')
 
   return (
     <>
